requirements.txt
 opencv-python==4.11.0.86
networkx==3.4.2
transformers~=4.48.3
torch~=2.6.0
pillow~=11.1.0
numpy~=2.2.3
scikit-learn~=1.6.1